Move away jinja filters
